RajJarvis-Github/requirements.txt
pywebview
SpeechRecognition
pyttsx3
pyaudio
